requirements.txt
flask
requests
python-dotenv
pytest
beautifulsoup4
pytest-cov
groq
google-cloud-storage
google-cloud-aiplatform
google-genai
supabase
supabase
Flask-Bcrypt
Flask-Limiter
Flask-Cors
Pillow
